Refactoring reporting repo structure
