+++ src/adaptiq/core/pipelines/pre_run/tools/prompt_consulting.py
@@ -94,7 +94,7 @@ class PromptConsulting:
         except Exception as e:
             raise ValueError(
                 f"Failed to parse LLM response as JSON: {e}\n\nResponse:\n{response.content}"
-            )
+            ) from e
 
         return result
