+++ src/adaptiq/core/reporting/aggregation/__init__.py
@@ -1,0 +1,3 @@
+from .aggregator import Aggregator
+
+__all__ = ["Aggregator"]
